refactor(config): isolate runtime transport policy interpretation
